Move OptionConverters to javaapi
